refactor(orchestration): extract shared application services and models
